requirements.txt
flask
pillow
numpy
requests
scipy